requirements.txt
# requirements.txt
torch
torchvision
transformers
accelerate
opencv-contrib-python
supervision
Pillow
tqdm
numpy